+ 改名功能

+++ tests/player/item/test_player_item_component.cpp
@@ -6,6 +6,7 @@
 #include "player/item/player_item_component.h"
 #include "player/mock/mock_database_pool.h"
 #include "player/mock/mock_player_state.h"
+#include "config/luban/generated/schema.h"
 
 #include <boost/asio/io_context.hpp>
 #include <boost/test/unit_test.hpp>
@@ -33,21 +34,21 @@ namespace
 
             auto item_config1 = std::make_shared<celeritas::config::container::item_config>();
             item_config1->itemTemplateId = 1001;
-            item_config1->itemType = static_cast<int>(celeritas::config::item_type::custom);
+            item_config1->itemType = celeritas::config::item_type::custom;
             item_config1->stacked = 100;
             item_config1->squares = false;
             dataMap->emplace(item_config1->itemTemplateId, item_config1);
 
             auto item_config2 = std::make_shared<celeritas::config::container::item_config>();
             item_config2->itemTemplateId = 1002;
-            item_config2->itemType = static_cast<int>(celeritas::config::item_type::equipment);
+            item_config2->itemType = celeritas::config::item_type::equipment;
             item_config2->stacked = 1;
             item_config2->squares = false;
             dataMap->emplace(item_config2->itemTemplateId, item_config2);
 
             auto item_config3 = std::make_shared<celeritas::config::container::item_config>();
             item_config3->itemTemplateId = 1003;
-            item_config3->itemType = static_cast<int>(celeritas::config::item_type::equipment);
+            item_config3->itemType = celeritas::config::item_type::equipment;
             item_config3->stacked = 1;
             item_config3->squares = false;
             dataMap->emplace(item_config3->itemTemplateId, item_config3);
